+++ loaders/target-gcs/target_gcs/paths/_partitioned/hive.py
@@ -8,7 +8,7 @@ from .validators import is_date_field
 def get_hive_path_generator(
     partition_fields: list[str], schema: dict[str, Any]
 ) -> list[tuple[str, Callable[[str, str], str]]]:
-    hive_path = []
+    hive_path: list[tuple[str, Callable[[str, str], str]]] = []
     for partition_field in partition_fields:
         if is_date_field(schema.get("properties", {}).get(partition_field)):
             hive_path.append((partition_field, date_as_partition))